requirements.txt
fastapi
uvicorn[standard]
requests
numpy
opencv-python
pillow
easyocr
pytesseract
python-multipart
pydantic
pandas
rapidfuzz
pdf2image
opencv-contrib-python